try to enable java tests in travis
