+++ java-package/brainflow/src/main/java/brainflow/examples/BrainFlowGetData.java
@@ -1,4 +1,4 @@
-package brainflow_test;
+package brainflow.examples;
 
 import java.util.Arrays;
