+++ java-package/brainflow/src/main/java/brainflow/examples/Downsampling.java
@@ -1,46 +1,44 @@
-package brainflow_test;
+package brainflow.examples;
 
 import java.util.Arrays;
 
+import brainflow.AggOperations;
 import brainflow.BoardIds;
 import brainflow.BoardShim;
 import brainflow.BrainFlowInputParams;
 import brainflow.DataFilter;
 import brainflow.LogLevels;
 
-public class Serialization
+public class Downsampling
 {
 
     public static void main (String[] args) throws Exception
     {
-        // use Synthetic board for demo
+        // use synthetic board for demo
         BoardShim.enable_board_logger ();
         BrainFlowInputParams params = new BrainFlowInputParams ();
         int board_id = BoardIds.SYNTHETIC_BOARD.get_code ();
-        BoardShim board_shim = new BoardShim (board_id, params);
 
+        BoardShim board_shim = new BoardShim (board_id, params);
         board_shim.prepare_session ();
         board_shim.start_stream (3600);
         BoardShim.log_message (LogLevels.LEVEL_INFO.get_code (), "Start sleeping in the main thread");
         Thread.sleep (5000);
         board_shim.stop_stream ();
         System.out.println (board_shim.get_board_data_count ());
-        int num_rows = BoardShim.get_num_rows (board_id);
         double[][] data = board_shim.get_current_board_data (30);
-        for (int i = 0; i < num_rows; i++)
-        {
-            System.out.println (Arrays.toString (data[i]));
-        }
         board_shim.release_session ();
 
-        // demo for serialization
-        DataFilter.write_file (data, "test.csv", "w");
-        double[][] restored_data = DataFilter.read_file ("test.csv");
-        System.out.println ("After Serialization:");
-        for (int i = 0; i < num_rows; i++)
+        int[] eeg_channels = BoardShim.get_eeg_channels (board_id);
+        for (int i = 0; i < eeg_channels.length; i++)
         {
-            System.out.println (Arrays.toString (restored_data[i]));
+            System.out.println ("Original data:");
+            System.out.println (Arrays.toString (data[i]));
+            // keep each second element, you can use MEAN and MEDIAN as well
+            double[] downsampled_data = DataFilter.perform_downsampling (data[eeg_channels[i]], 2,
+                    AggOperations.EACH.get_code ());
+            System.out.println ("Downsampled data:");
+            System.out.println (Arrays.toString (downsampled_data));
         }
     }
-
 }
